pyproject.toml
[tool.poetry]
name = "airbnb-backend"
version = "0.1.0"
description = ""
authors = ["'heokyoungmin' <'stalvede0421@gmail.com'>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
black = "^22.10.0"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.149.2"}
strawberry-graphql-django = "^0.8.2"
PyJWT = "^2.6.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
